Introduce XTRIM by RANGE.
This allows streams to be trimmed by an inclusive range of IDs. Like
XTRIM with MAXLEN, you can perform approximate trimming which will only
drops nodes rather than deleting entries which is more efficient.

Users should be judicial with the use of this strategy since there is no
guarantee on runtime, especially with varying insertion rate.